puts2

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -9,7 +9,7 @@
 void puts2(char *str)
 {
 	int i = 0;
-	size_t length = strlen(str);
+	int length = strlen(str);
 
 	while (str[i] != '\0' && i < length)
 	{
